opitmized_RF_function

opitmized_RF_function

optimized_RF_function(
  data,
  species,
  response_data,
  trait_data,
  phylo_data = NULL,
  classification,
  weight = NULL,
  mtry_frac = NULL,
  min.node.size = NULL,
  sample.fraction = NULL,
  ntrees = NULL,
  PEMs = NULL,
  cores = 1
)

Arguments

data

A species_traits dataframe containing species traits and eigen values if the user use phylogenetic data

species

The name of the column corresponding to the species names in the @param data

response_data

The name of the column corresponding to the response in the @param data

trait_data

A dataframe containing only species trait data

phylo_data

By default, phylo_data = NULL, which return an expand grid without PEMs

classification

Can be eiter TRUE or FALSE. If TRUE, you want to perform a classification analyses, if FALSE, you want to perform a regression analyses

weight

A list containing vector as the same legnth of row in the data

mtry_frac

By default, NULL, return a list of default values of mtry. Else, you define your own range of mtry value you want to optimize in a vector

min.node.size

By default, NULL, return a list of default values of minimum node size. Else, you define your own range of minimum node size value you want to optimize in a vector

sample.fraction

By default, NULL, return a list of default values of fraction of observations to sample. Else, you define your own range of observations to sample value you want to optimize in a vector

ntrees

By default, NULL, return a list of default values of number of trees. Else, you define your own range of number of trees you want to optimize in a vector

PEMs

ggg

cores

Define the number of cores you want to use. By default, the number of cores is set to 1. For Windows user, keep number of cores to 1, or switch to linux/Mac OS.

Value

Return a dataframe

Examples

if (FALSE) {
library(predipreda)
data(dataset)
species <- colnames(dataset$Species_ID)
response_data <- colnames(dataset$Response)
trait_data <- dataset[,c(2:5)]
results <- optimized_RF_function(dataset,species,response_data,trait_data,phylo_data = TRUE, classification == FALSE)
}